feat(GER-45-configurar-styled-components): Adição das configurações do styled-components, adequação das screens para o novo estilo de estilização
