+++ package.json
@@ -17,11 +17,16 @@
     "react-native": "0.72.6",
     "react-native-safe-area-context": "4.6.3",
     "react-native-screens": "~3.22.0",
-    "react-native-svg": "13.9.0"
+    "react-native-svg": "13.9.0",
+    "styled-components": "^6.1.6"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
+    "babel-plugin-styled-components": "^2.1.4",
     "react-native-svg-transformer": "^1.3.0"
   },
-  "private": true
+  "private": true,
+  "resolutions": {
+    "styled-components": "^5"
+  }
 }
