+++ src/routes/AppRoutes.js
@@ -1,11 +1,14 @@
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import Home from "../screens/Home";
-import Settings from "../screens/Settings";
+import Home from "../screens/Home/Home";
+import Settings from "../screens/Settings/Settings";
 
 const AppRoutes = () => {
   const Stack = createNativeStackNavigator();
   return (
-    <Stack.Navigator>
+    <Stack.Navigator
+      initialRouteName="Home"
+      screenOptions={{ headerShown: false }}
+    >
       <Stack.Screen name="Home" component={Home} />
       <Stack.Screen name="Settings" component={Settings} />
     </Stack.Navigator>
