--- src/screens/Home.js
@@ -1,21 +1,0 @@
-import { View, Text, StyleSheet } from "react-native";
-import React from "react";
-
-const Home = () => {
-  return (
-    <View style={styles.container}>
-      <Text>Home</Text>
-    </View>
-  );
-};
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-});
-
-export default Home;
