+++ src/screens/Home/Home.js
@@ -1,0 +1,13 @@
+import { Text } from "react-native";
+import React from "react";
+import { Container } from "./styles";
+
+const Home = () => {
+  return (
+    <Container>
+      <Text>Home</Text>
+    </Container>
+  );
+};
+
+export default Home;
