+++ src/screens/Settings/Settings.js
@@ -1,0 +1,13 @@
+import { Text } from "react-native";
+import React from "react";
+import { Container } from "./styles";
+
+const Settings = () => {
+  return (
+    <Container>
+      <Text>Settings</Text>
+    </Container>
+  );
+};
+
+export default Settings;
