+++ src/screens/Settings/styles.js
@@ -1,0 +1,8 @@
+import styled from "styled-components/native";
+
+export const Container = styled.View`
+  display: flex;
+  flex: 1;
+  align-items: center;
+  justify-content: center;
+`;
