--- src/screens/SignUp.js
@@ -1,19 +1,0 @@
-import { View, Text, StyleSheet } from "react-native";
-import React from "react";
-
-const SignUp = () => {
-  return (
-    <View style={styles.container}>
-      <Text>SignUp</Text>
-    </View>
-  );
-};
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-});
-export default SignUp;
