+++ src/screens/SignUp/SignUp.js
@@ -1,0 +1,13 @@
+import { Text } from "react-native";
+import React from "react";
+import { Container } from "./styles";
+
+const SignUp = () => {
+  return (
+    <Container>
+      <Text>SignUp</Text>
+    </Container>
+  );
+};
+
+export default SignUp;
